fix: move retrieval functions to separate file
